// Frontend/ClientApp/app/hoshin-quality/corrective-actions/models/Task.ts
export class Task{
    taskID: number;
    taskStateID: number;
    entityID: number;
    description: string;
    implementationPlannedDate: Date;
    requireEvidence: boolean;
    firstname: string;
    surname: string;
    plantID: number;
    sectorID: number;
    acApporverOrResponsibleSGCCanEdit: boolean;
    responsibleUserID: string;

}